config stuff

+++ src/main/java/frc/robot/constants/swerve/moduleConfigs/SwerveModuleGeneralConfigBase.java
@@ -61,5 +61,5 @@ public abstract class SwerveModuleGeneralConfigBase {
     public abstract FeedbackSensorSourceValue getSteerCancoderFeedbackSensorSource();
 
     public abstract SensorDirectionValue getCancoderSensorDirection();
-    public abstract int getCancoderAbsoluteSensorDiscontinuityPoint();
+    public abstract double getCancoderAbsoluteSensorDiscontinuityPoint();
 }
